requirements.txt
python-dotenv
requests
beautifulsoup4
openai
flask